+++ mcp_tools/exploit_framework/msfvenom.py
@@ -1,0 +1,36 @@
+# mcp_tools/exploit_framework/msfvenom.py
+
+from typing import Dict, Any
+
+def register_msfvenom(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def msfvenom_generate(payload: str, format_type: str = "", output_file: str = "", encoder: str = "", iterations: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute MSFVenom for payload generation with enhanced logging.
+
+        Args:
+            payload: The payload to generate
+            format_type: Output format (exe, elf, raw, etc.)
+            output_file: Output file path
+            encoder: Encoder to use
+            iterations: Number of encoding iterations
+            additional_args: Additional MSFVenom arguments
+
+        Returns:
+            Payload generation results
+        """
+        data = {
+            "payload": payload,
+            "format": format_type,
+            "output_file": output_file,
+            "encoder": encoder,
+            "iterations": iterations,
+            "additional_args": additional_args
+        }
+        logger.info(f"🚀 Starting MSFVenom payload generation: {payload}")
+        result = hexstrike_client.safe_post("api/tools/msfvenom", data)
+        if result.get("success"):
+            logger.info(f"✅ MSFVenom payload generated")
+        else:
+            logger.error(f"❌ MSFVenom payload generation failed")
+        return result
